refactor: Callout 디자인 토큰 기반 수정 + 파일 경로 수정
