+++ frontend/src/Pages/AdverseReliefPage.tsx
@@ -1,4 +1,4 @@
-import Callout from '@/components/Callout';
+import Callout from '@/components/ui/Callout';
 import PageLayout from '@/components/PageLayout';
 import ReliefProcess from '@/components/ReliefProcess';
